+++ resources/views/medical_records/show.blade.php
@@ -29,9 +29,11 @@
                 {!! link_to_with_icon('fas fa-arrow-left fa-2x', 'patients.medical_records.index', $patient, 'Back to Medical Records', 'has-text-info') !!}
             </p>
 
-            <p class="control">
-                {!! delete_link_with_icon('fas fa-trash-alt fa-2x', 'patients.medical_records.destroy', [$patient, $medical_record], 'Delete Medical Record', 'has-text-danger') !!}
-            </p>
+            @can ('delete', $medical_record)
+                <p class="control">
+                    {!! delete_link_with_icon('fas fa-trash-alt fa-2x', 'patients.medical_records.destroy', [$patient, $medical_record], 'Delete Medical Record', 'has-text-danger') !!}
+                </p>
+            @endcan
         </div>
     </div>
